fix(): angular variable undefined when window.angular is used
There was a bug in the code where we accidently created an angular
variable inside of the function scope, this was preventing window.angular
or window.showdown from working. (But it was working fine with node
and require). I wrapped the file in a IIFE that injects angular and showdown.
This solves the problem.

No breaking changes.